feat: Add meal editing functionality to DailyPage with reusable form component
- Extract meal entry form from MealPage into reusable MealEntryForm component
- Add support for both 'create' and 'edit' modes in MealEntryForm
- Implement meal editing modal on DailyPage with edit button next to delete
- Add proper form pre-population for edit mode with existing meal data
- Include food intelligence data loading in DailyPage for form auto-completion
- Optimize Edit/Delete button spacing with reduced padding (w-7/w-8 vs w-8/w-9)
- Maintain all existing functionality: validation, sync handling, food intelligence
- Ensure responsive design and touch-friendly interactions on mobile

Files changed:
- src/components/MealEntryForm.tsx (new): Extracted reusable form component
- src/pages/MealPage.tsx: Refactored to use new MealEntryForm component
- src/pages/DailyPage.tsx: Added edit functionality with modal dialog

This enables users to edit meal entries directly from the daily summary view
without navigating away from the page, improving the overall user experience.
